+++ src/GovUk.Frontend.AspNetCore/TagHelpers/FooterTagHelper.cs
@@ -1,6 +1,7 @@
 using System.Text.Encodings.Web;
 using GovUk.Frontend.AspNetCore.ComponentGeneration;
 using Microsoft.AspNetCore.Razor.TagHelpers;
+using Microsoft.Extensions.Options;
 
 namespace GovUk.Frontend.AspNetCore.TagHelpers;
 
@@ -21,17 +22,23 @@ public class FooterTagHelper : TagHelper
     private const string ContainerClassAttributeName = "container-class";
 
     private readonly IComponentGenerator _componentGenerator;
+    private readonly IOptions<GovUkFrontendOptions> _optionsAccessor;
     private readonly HtmlEncoder _encoder;
 
     /// <summary>
     /// Creates a new <see cref="FooterTagHelper"/>.
     /// </summary>
-    public FooterTagHelper(IComponentGenerator componentGenerator, HtmlEncoder encoder)
+    public FooterTagHelper(
+        IComponentGenerator componentGenerator,
+        IOptions<GovUkFrontendOptions> optionsAccessor,
+        HtmlEncoder encoder)
     {
         ArgumentNullException.ThrowIfNull(componentGenerator);
+        ArgumentNullException.ThrowIfNull(optionsAccessor);
         ArgumentNullException.ThrowIfNull(encoder);
 
         _componentGenerator = componentGenerator;
+        _optionsAccessor = optionsAccessor;
         _encoder = encoder;
     }
 
@@ -65,7 +72,8 @@ public class FooterTagHelper : TagHelper
             Copyright = footerContext.Copyright?.Options,
             ContainerClasses = ContainerClass,
             Classes = classes,
-            Attributes = attributes
+            Attributes = attributes,
+            Rebrand = _optionsAccessor.Value.Rebrand
         });
 
         output.ApplyComponentHtml(component, _encoder);