+++ tests/GovUk.Frontend.AspNetCore.IntegrationTests/StartupFilterTests.cs
@@ -34,6 +34,9 @@ public class StartupFilterTests
     [InlineData("images/favicon.ico")]
     [InlineData("images/govuk-crest.svg")]
     [InlineData("fonts/bold-affa96571d-v2.woff")]
+    [InlineData("manifest.json")]
+    [InlineData("rebrand/images/favicon.ico")]
+    [InlineData("rebrand/manifest.json")]
     public async Task HostsStaticAssets(string fileName)
     {
         // Arrange